requirements.txt
torch==1.4.0
tokenizers==0.5.2 
transformers==4.2.2
seaborn==0.10.1
absl-py==0.9.0
loguru==0.5.1
scikit-learn
tensorboard
git+git://github.com/erikavaris/tokenizer.git
